requirements.txt
aiomysql==0.1.1
amqp==5.1.1
anyio==3.6.2
arrow==1.2.3
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==40.0.1
dnspython==2.3.0
eventlet==0.33.3
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
numpy==1.24.2
pandas==1.5.3
prompt-toolkit==3.0.38
protobuf==3.20.3
pycparser==2.21
pydantic==1.10.7
PyMySQL==1.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytimeparse==1.1.8
pytz==2023.2
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.8
starlette==0.26.1
typer==0.7.0
typing-extensions==4.5.0
uvicorn==0.21.1
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4